% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{augment.oomlm}
\alias{augment.oomlm}
\title{augment data with prediction, std error, and residuals}
\usage{
\method{augment}{oomlm}(x, data, std_error = FALSE,
  interval = "confidence", ...)
}
\arguments{
\item{x}{\code{oomlm()} model}

\item{data}{\code{tibble()} or other data source}

\item{std_error}{calculate standard error of prediction}

\item{interval}{interval type to return}

\item{...}{ignored}
}
\description{
augment data with prediction, std error, and residuals
}
\seealso{
\code{\link[=augment]{augment()}}, \code{\link[stats:predict.lm]{stats::predict.lm()}}
}
