% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounded_qr.R
\name{coef.Rcpp_BoundedQr}
\alias{coef.Rcpp_BoundedQr}
\title{wrapper for `BoundedQr` method `betas`}
\usage{
\method{coef}{Rcpp_BoundedQr}(qr, nvar = NULL, ...)
}
\arguments{
\item{qr}{BoundedQr object}

\item{nvar}{int number of coefficients to return}

\item{...}{ignored}
}
\description{
wrapper for `BoundedQr` method `betas`
}
\keyword{internal}
