% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_model.R
\name{deparse2}
\alias{deparse2}
\title{lifted as-is from \code{stats}}
\usage{
deparse2(x)
}
\arguments{
\item{x}{variables extracted from \code{\link[=terms]{terms()}} object}
}
\description{
lifted as-is from \code{stats}
}
\keyword{internal}
