% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomlm_summary.R
\name{dispersion_oomlm}
\alias{dispersion_oomlm}
\title{Internal. Calculate dispersion of oomlm model.}
\usage{
dispersion_oomlm(x, ...)
}
\arguments{
\item{x}{\code{oomlm} object}

\item{...}{ignored}
}
\description{
Internal. Calculate dispersion of oomlm model.
}
\keyword{internal}
