% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm.R
\name{end_reweight}
\alias{end_reweight}
\title{Perform end of IWLS reweight iteration}
\usage{
end_reweight(obj, tolerance = 1e-07)
}
\arguments{
\item{obj}{\code{oomglm} model.}

\item{tolerance}{Tolerance for change in coefficient as a multiple
of standard error.}
}
\description{
Update variables used for IWLS calculation, increment \code{iter}, and
determine if model has converged
}
