% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm.R
\name{end_weight}
\alias{end_weight}
\title{Complete model for IRLS iteration}
\usage{
end_weight(object, tolerance = 1e-07)
}
\arguments{
\item{object}{\code{oomglm} model.}

\item{tolerance}{Tolerance for change in coefficient as a multiple
of standard error.}
}
\description{
Update variables used for IRLS calculation, increment \code{iter}, and
determine if model has converged
}
