% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oom_tidy.R
\name{fit.oomglm}
\alias{fit.oomglm}
\title{Fit \code{\link[=oomglm]{oomglm()}} model via Iteratively Reweighted Least Squares (IRLS).}
\usage{
\method{fit}{oomglm}(object, data, times = 2L, tolerance = 1e-08, ...)
}
\arguments{
\item{object}{\code{\link[=oomglm]{oomglm()}} model.}

\item{data}{\code{\link[=oomdata_tbl]{oomdata_tbl()}}, \code{\link[=oomdata_dbi]{oomdata_dbi()}}, \code{\link[=oomdata_con]{oomdata_con()}},
\code{\link[=tibble]{tibble()}}, \code{\link[=data.frame]{data.frame()}}, or \code{\link[=list]{list()}} of observations to fit}

\item{times}{Maximum number of IRLS iterations to perform. Will
stop iterating if model converges before \code{times} iterations.}

\item{tolerance}{Tolerance used to determine convergence. Represents
change in coefficient as a multiple of standard error.}

\item{...}{ignored}
}
\value{
\code{\link[=oomglm]{oomglm()}} object after performing \code{times} IRLS iterations on
\code{data}.
}
\description{
Fit \code{\link[=oomglm]{oomglm()}} model via Iteratively Reweighted Least Squares (IRLS).
}
\seealso{
\code{\link[=oomglm]{oomglm()}}
}
