% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oom_tidy.R, R/oomlm.R
\name{fit.oomlm}
\alias{fit.oomlm}
\alias{update.oomlm}
\title{Fit \code{oomlm()} model to additonal observations}
\usage{
\method{fit}{oomlm}(object, data, ...)

\method{update}{oomlm}(object, data, ...)
}
\arguments{
\item{object}{\code{\link[=oomlm]{oomlm()}} model to be updated}

\item{data}{an optional \code{\link[=oomdata_tbl]{oomdata_tbl()}}, \code{\link[=oomdata_dbi]{oomdata_dbi()}}, \code{\link[=oomdata_con]{oomdata_con()}},
\code{\link[=tibble]{tibble()}}, \code{\link[=data.frame]{data.frame()}}, or \code{\link[=list]{list()}} of observations to fit}

\item{...}{ignored}
}
\description{
Update ploom model fit with new data.
}
\seealso{
\code{\link[=oomlm]{oomlm()}}
}
