% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_data.R
\name{generate_data}
\alias{generate_data}
\title{Intenal. For generating GLM data used in tests and vignettes.}
\usage{
generate_data(betas, nrows)
}
\arguments{
\item{betas}{coefficients}

\item{nrows}{number of observations to generate}
}
\description{
Create data with generalized linear relationship.
}
\keyword{internal}
