% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomdata.R
\name{get_oomdata}
\alias{get_oomdata}
\title{Internal. Extract information from \code{oomdata} function}
\usage{
get_oomdata(obj)
}
\arguments{
\item{obj}{An \code{oomdata} function.}
}
\description{
Internal. Extract information from \code{oomdata} function
}
\keyword{internal}
