% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{glance.oomlm}
\alias{glance.oomlm}
\alias{glance.summary.oomlm}
\alias{glance.oomglm}
\title{Returns a tibble with exactly one row of goodness of fitness measures
and related statistics.}
\usage{
\method{glance}{oomlm}(x, ...)

\method{glance}{summary.oomlm}(x, ...)

\method{glance}{oomglm}(x, ...)
}
\arguments{
\item{x}{\code{oomlm} model object}

\item{...}{ignored.}
}
\value{
A one-row \link[tibble:tibble]{tibble::tibble} with columns:

\item{r.squared}{The percent of variance explained by the model}
\item{adj.r.squared}{r.squared adjusted based on the degrees of freedom}
\item{sigma}{The square root of the estimated residual variance}
\item{statistic}{F-statistic}
\item{p.value}{p-value from the F test, describing whether the full
regression is significant}
\item{df}{Degrees of freedom used by the coefficients}
\item{logLik}{the data's log-likelihood under the model}
\item{AIC}{the Akaike Information Criterion}
\item{BIC}{the Bayesian Information Criterion}
\item{deviance}{deviance}
\item{df.residual}{residual degrees of freedom}
}
\description{
Returns a tibble with exactly one row of goodness of fitness measures
and related statistics.
}
\seealso{
\code{\link[=glance]{glance()}}
}
