% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm_robust.R
\name{init_oomglm_robust}
\alias{init_oomglm_robust}
\title{Initialize Generalized Linear Regression Model
with robust standard errors}
\usage{
init_oomglm_robust(formula, family, weights, start, se_type)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted of class
\code{formula}.}

\item{family}{a \code{\link[=family]{family()}} object.}

\item{weights}{a one-sided, single term \code{formula} specifying weights.}

\item{start}{starting values for the parameters in the linear predictor.}

\item{se_type}{type of robust standard errors}
}
\description{
Initialize Generalized Linear Regression Model
with robust standard errors
}
\keyword{internal}
