% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomlm.R
\name{init_oomlm}
\alias{init_oomlm}
\title{initialize generalized linear model}
\usage{
init_oomlm(formula, weights = NULL)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted
of class \code{formula}.}

\item{weights}{A one-sided, single term \code{formula} specifying weights.}
}
\description{
initialize generalized linear model
}
\keyword{internal}
