% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomlm_robust.R
\name{init_oomlm_robust}
\alias{init_oomlm_robust}
\title{Internal. Initialize generalized linear model with robust standard errors}
\usage{
init_oomlm_robust(formula, weights, se_type)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted
of class \code{formula}.}

\item{weights}{A one-sided, single term \code{formula} specifying weights.}

\item{se_type}{Method for robust standard error estimation.}
}
\description{
Internal. Initialize generalized linear model with robust standard errors
}
\keyword{internal}
