% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm.R
\name{init_reweight}
\alias{init_reweight}
\title{Prepare model for IWLS iteration}
\usage{
init_reweight(obj)
}
\arguments{
\item{obj}{\code{oomglm} model.}
}
\description{
Reset variables used for IWLS calculation
}
\examples{
x    <- oomglm(mpg ~ cyl + disp)
feed <- oomfeed(mtcars, chunksize = 10)

# manually execute an IWLS reweight iteration
x <- init_reweight(x)
x <- update(x, feed)
x <- end_reweight(x)

}
