% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm.R, R/oomglm_irls.R
\name{init_weight}
\alias{init_weight}
\alias{weight}
\alias{end_weight}
\title{Reweighting functions that comprise an IRLS iteration}
\usage{
init_weight(object)

weight(object, data)

end_weight(object, tolerance = 1e-08)
}
\arguments{
\item{object}{\code{oomglm} model.}

\item{data}{an optional \code{oomdata_tbl}, \code{oomdata_dbi}, \code{oomdata_con},
\code{tibble}, \code{data.frame}, or \code{list} of observations to fit}

\item{tolerance}{Tolerance used to determine convergence. Represents
change in coefficient as a multiple of standard error.}
}
\value{
\code{oomglm} model
}
\description{
Reweighting functions that comprise an IRLS iteration
}
\details{
The IRLS process comprises iterative calls to \code{init_weight()}, \code{weight()}
and \code{end_weight()} which can be managed via a single call to \code{fit()}.

These comprising functions are exposed for lower level control of the
fitting process.
}
\examples{
\donttest{
# initiate `oomglm` model and `oomdata_tbl`
x <- oomglm(mpg ~ cyl + disp)
chunks <- oomdata_tbl(mtcars, chunk_size = 10)

# manually perform one round of IRLS and check results
x <- init_weight(x)
x <- weight(x, data = chunks)
x <- end_weight(x)

print(x$converged)
print(coef(x))

# manually perform an additional round of IRLS
x <- init_weight(x)
x <- weight(x, data = chunks)
x <- end_weight(x)

print(x$converged)
print(coef(x))

# the IRLS process as implemented by `fit()`
# is similar to the following
x <- oomglm(mpg ~ cyl + disp)
chunks <- oomdata_tbl(mtcars, chunk_size = 10)

while(!x$converged) {
  
  x <- init_weight(x)
  x <- weight(x, data = chunks)
  x <- end_weight(x) 
  
}

tidy(x)

}
}
\seealso{
\code{\link[=fit]{fit()}}, \code{\link[=oomglm]{oomglm()}}
}
