% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_data.R
\name{linear_params}
\alias{linear_params}
\title{Intenal. For generating GLM data used in tests and vignettes.}
\usage{
linear_params(p, betas_mu_lim = 0, betas_sd_lim = 1)
}
\arguments{
\item{p}{number of columns in X.  will be forced to the nearest
number divisible by two}

\item{betas_mu_lim}{mean of beta normal distributions
range between \code{-betas_mu_lim} and \code{betas_mu_lim}}

\item{betas_sd_lim}{standard deviation of beta normal distributions
range between 1 and \code{betas_sd_lim}}
}
\description{
Make a and B for y = a + BX + e.
}
\keyword{internal}
