% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_mfmm.R
\name{model_frame}
\alias{model_frame}
\title{Lifted from \code{stats}. modified to be faster and less robust}
\usage{
model_frame(formula, data = NULL)
}
\arguments{
\item{formula}{\code{\link[=terms]{terms()}} object}

\item{data}{data.frame or list}
}
\description{
Lifted from \code{stats}. modified to be faster and less robust
}
\keyword{internal}
