% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_mfmm.R
\name{model_frame_tibble}
\alias{model_frame_tibble}
\title{return model_frame value as tibble}
\usage{
model_frame_tibble(x, data)
}
\arguments{
\item{x}{\code{oomlm} or \code{oomglm} object}

\item{data}{tibble or data.frame}
}
\description{
return model_frame value as tibble
}
\keyword{internal}
