% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_model.R
\name{model_matrix}
\alias{model_matrix}
\title{Lifted from \code{stats}. modified to be faster and less robust}
\usage{
model_matrix(terms, data)
}
\arguments{
\item{data}{\code{\link[=data.frame]{data.frame()}} or \code{\link[=list]{list()}}}

\item{formula}{\code{\link[=terms]{terms()}} object}
}
\description{
Lifted from \code{stats}. modified to be faster and less robust
}
\keyword{internal}
