% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_model.R
\name{model_offset}
\alias{model_offset}
\title{Lifted from \code{stats}. modified to be faster and less robust}
\usage{
model_offset(x)
}
\arguments{
\item{x}{\code{\link[=formula]{formula()}} object}
}
\description{
Lifted from \code{stats}. modified to be faster and less robust
}
\keyword{internal}
