% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounded_qr.R
\name{new_bounded_qr}
\alias{new_bounded_qr}
\title{initialize new `BoundedQr` object}
\usage{
new_bounded_qr(np)
}
\arguments{
\item{np}{int number of independent parameters in model
including the intercept when model has intercept}
}
\description{
initialize new `BoundedQr` object
}
\keyword{internal}
