% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm.R
\name{ols_model}
\alias{ols_model}
\title{return true if model is gaussian family with identity link}
\usage{
ols_model(object)
}
\arguments{
\item{object}{\code{oomglm} model}
}
\description{
not currently in use
}
\keyword{internal}
