% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_lm.R
\name{online_lm}
\alias{online_lm}
\title{Linear model that uses only `p^2` memory for `p` variables.}
\usage{
online_lm(data, formula, weights = NULL, sandwich = FALSE)
}
\arguments{
\item{data}{an optional data frame, list or environment
(or object coercible by as.data.frame to a data frame)}

\item{formula}{A model formula: a symbolic description of the
model to be fitted.}

\item{weights}{A one-sided, single term formula specifying weights}

\item{sandwich}{TRUE to compute the Huber/White sandwich covariance matrix
(uses p^4 memory rather than p^2)}
}
\description{
Linear model that uses only `p^2` memory for `p` variables.
}
\details{
The model formula must not contain any data-dependent terms, as
  these will not be consistent when updated. Factors are permitted, but 
  the levels of the factor must be the same across all data chunks 
  (empty factor levels are ok). Offsets are allowed.
}
