% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oom_tidy.R
\name{oom_tidiers}
\alias{oom_tidiers}
\alias{tidy.oomlm}
\title{Tidy an oomlm model}
\usage{
\method{tidy}{oomlm}(x, ...)
}
\arguments{
\item{x}{oomlm object}

\item{...}{ignored}
}
\value{
A data.frame with columns for coefficient names, estimates, standard
errors, confidence intervals, p-values, degrees of freedom, and the
name of the outcome variable
}
\description{
Tidy an oomlm model
}
