% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomdata.R
\name{oomdata_dbi}
\alias{oomdata_dbi}
\alias{oomdata_tbl}
\alias{oomdata_con}
\title{Iterate over \code{tibble()}, \code{DBIResult} or \code{connection()} in chunks.}
\usage{
oomdata_tbl(data, chunk_size, ...)

oomdata_dbi(data, chunk_size, ...)

oomdata_con(data, chunk_size, header = TRUE, col_names = NULL, ...)
}
\arguments{
\item{data}{\code{\link[=tibble]{tibble()}}, \code{DBIResult}, or \code{\link[=connection]{connection()}} object}

\item{chunk_size}{number of chunks to return with each iteration}

\item{...}{ignored}

\item{header}{logical, when TRUE colnames are determined from first row
of connection. if FALSE \code{col_names} must be provided.}

\item{col_names}{character vector. override \code{col_names} in first
row of connection when \code{header}is TRUE, or provide colnames
if \code{header} is FALSE.}
}
\description{
Returns a function that repeats the following cycle:
iteratively return \code{chunk_size} number of rows from \code{data} until data is exhausted;
then return \code{NULL} once.
}
\details{
\code{oomdata_*} functions are closures that return functions that
iteratively return \code{chunk_size} number of rows from \code{data} until all
rows have been returned. They will then return \code{NULL} once. They repeat
this cycle ad-infinitum.
}
