% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.R
\name{oomfeed}
\alias{oomfeed}
\alias{oomfeed.data.frame}
\alias{oomfeed.connection}
\alias{oomfeed.DBIResult}
\title{Repeatedly iterate over \code{data.frame} or \code{connection} in chunks.}
\usage{
oomfeed(data, chunk_size, ...)

oomfeed.data.frame(data, chunk_size, ...)

oomfeed.connection(data, chunk_size, header = TRUE, col_names = NULL,
  ...)

oomfeed.DBIResult(data, chunk_size, ...)
}
\arguments{
\item{data}{\code{data.frame}, \code{file}, \code{gzfile}, or \code{url}connection`}

\item{chunk_size}{number of chunks to return with each iteration. overrides
\code{nrow} parameter of \code{read.table} when \code{data} is a connection.}

\item{...}{passed through to \code{read.table}. note that \code{nrow}, and
\code{col.names} are ignored in favor of \code{chunk_size} and \code{col_names}}

\item{header}{logical, when TRUE colnames are determined from first row
of connection. if FALSE \code{col_names} must be provided.}

\item{col_names, }{character vector. override \code{col_names} in first
row of connection when \code{header}is TRUE, or provide colnames
if \code{header} is FALSE.}
}
\description{
Returns a function that repeats the following cycle:
iteratively return \code{chunk_size} number of rows from \code{data} until data is exhausted;
then return \code{NULL} once.
}
\details{
\code{oomfeed} is a closure that creates a function for
returning \code{chunk_size} number of rows from \code{data} until all
rows have been returned.  It will then return \code{NULL} once. It will repeat
this cycle ad-infinitum. When data is a \code{connection} it will recreate and open
the connection with each cycle.  It reads from \code{connection} objects using
\code{read.table}.  Additional parameters for \code{read.table} should be passed
in via \code{...}.
}
