% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oom_predict.R
\name{oomglm_residuals}
\alias{oomglm_residuals}
\title{calculate residuals for oomglm}
\usage{
oomglm_residuals(object, data, type = c("deviance", "pearson",
  "response", "working"), as_function = FALSE)
}
\arguments{
\item{object}{\code{oomlm} model}

\item{data}{data to calculate residuals}

\item{type}{residual calculation method}

\item{as_function}{if TRUE a function with only a \code{data} argument is returned
for subsequent residual calculations}
}
\description{
calculate residuals for oomglm
}
