% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm_robust.R
\name{oomglm_robust}
\alias{oomglm_robust}
\alias{vcov.oomglm_robust}
\title{Out of memory Generalized Linear model with robust standard errors}
\usage{
oomglm_robust(formula, family = gaussian(), weights = NULL,
  start = NULL, se_type = "H1", ...)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted of class
\code{\link[=formula]{formula()}}}

\item{family}{A \code{\link[=family]{family()}} object}

\item{weights}{a one-sided, single term \code{\link[=formula]{formula()}} specifying weights}

\item{start}{starting values for the parameters in the linear predictor}

\item{se_type}{string indicating what \code{se} type to use: "HC0", "HC1"
"stata", or "classical"}

\item{...}{ignored}

\item{data}{an optional \code{\link[=oomdata_tbl]{oomdata_tbl()}}, \code{\link[=oomdata_dbi]{oomdata_dbi()}}, \code{\link[=oomdata_con]{oomdata_con()}},
\code{\link[=tibble]{tibble()}}, \code{\link[=data.frame]{data.frame()}}, or \code{\link[=list]{list()}} of observations to fit}
}
\value{
It is up to the user to know when fitting is complete.
Therefore, only basic model characteristics are provided as values.
Statistics are available on demand via summary and extractor functions.

\item{converged}{Indicates if the IRLS algorithm has converged.}
\item{iter}{The number of iterations of IRLS performed.}
\item{n}{The number observations processed per round of IRLS.}
\item{df.residual}{The residual degrees of freedom.}
\item{df.null}{The residual degrees of freedom.}
\item{formula}{the \code{\link[=formula]{formula()}} object specifying the linear model.}
\item{family}{a \code{\link[=family]{family()}} object describing the error distribution
and link function used in the model.}
\item{terms}{The \code{\link[=terms]{terms()}} object used.}
\item{weights}{The weights \code{\link[=formula]{formula()}} provided to the model.}
\item{call}{The matched call.}
}
\description{
Perform linear regression via Alan Miller's bounded memory QR
factorization algorithm and estimate robust standard errors.
}
\details{
The provided \code{\link[=formula]{formula()}} must not contain any data-dependent terms to ensure
consistency across calls to \code{\link[=update]{update()}}. Factors are permitted, but the
levels of the factor must be the same across all data chunks. Empty factor
levels are accepted.
}
\seealso{
\code{\link[=oomglm]{oomglm()}}
}
