% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oom_predict.R
\name{oomlm_residuals}
\alias{oomlm_residuals}
\title{calculate residuals for oomlm}
\usage{
oomlm_residuals(object, data, as_function = FALSE)
}
\arguments{
\item{object}{\code{oomlm} model}

\item{data}{data to calculate residuals}

\item{as_function}{if TRUE a function with only a \code{data} argument is returned
for subsequent residual calculations}
}
\description{
calculate residuals for oomlm
}
