% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomlm_robust.R
\name{oomlm_robust}
\alias{oomlm_robust}
\alias{vcov.oomlm_robust}
\title{Out of memory Linear model with robust standard errors}
\usage{
oomlm_robust(formula, data = NULL, weights = NULL, se_type = "HC1",
  ...)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted of class
\code{\link[=formula]{formula()}}.}

\item{data}{an optional \code{\link[=oomdata_tbl]{oomdata_tbl()}}, \code{\link[=oomdata_dbi]{oomdata_dbi()}}, \code{\link[=oomdata_con]{oomdata_con()}},
\code{\link[=tibble]{tibble()}}, \code{\link[=data.frame]{data.frame()}}, or \code{\link[=list]{list()}} of observations to fit}

\item{weights}{a one-sided, single term \code{\link[=formula]{formula()}} specifying weights.}

\item{se_type}{string indicating what \code{se} type to use: "HC0", "HC1"
"stata", or "classical"}

\item{...}{ignored.}
}
\value{
A \code{\link[=oomlm_robust]{oomlm_robust()}} model is perpetually in an \emph{in-progress} state. It is up
to the user to know when fitting is complete. Therefore, only basic
model characteristics are provided as values. Statistics are available on
demand via \code{summary} and \code{extractor} functions.

\item{n}{the number of observations processed.}
\item{df.residual}{the residual degrees of freedom.}
\item{formula}{the \code{\link[=formula]{formula()}} object specifying the linear model.}
\item{terms}{the \code{\link[=terms]{terms()}} object specifying the terms of the linear model.}
\item{weights}{a one-sided, single term \code{\link[=formula]{formula()}} specifying weights.}
\item{call}{the matched call.}
}
\description{
Perform linear regression via Alan Miller's bounded memory QR
factorization algorithm and estimate robust standard errors.
}
\details{
The provided \code{\link[=formula]{formula()}} must not contain any data-dependent terms to ensure
consistency across calls to \code{\link[=update]{update()}}. Factors are permitted, but the
levels of the factor must be the same across all data chunks. Empty factor
levels are accepted.
}
\seealso{
\code{\link[=oomlm]{oomlm()}}
}
