% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_data.R
\name{params_as_tibble}
\alias{params_as_tibble}
\title{Intenal. For generating GLM data used in tests and vignettes.}
\usage{
params_as_tibble(params)
}
\arguments{
\item{params}{coefficient values}

\item{col_names}{names of the coefficients}
}
\description{
Transform param list into single row table
with alpha and B having names corresponding to data
}
\keyword{internal}
