\name{ploom-package}
\alias{ploom-package}
\alias{ploom}
\docType{package}
\title{
\packageTitle{ploom}
}
\description{
\packageDescription{ploom}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ploom}
\packageIndices{ploom}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ploom}

Maintainer: \packageMaintainer{ploom}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
