% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_oomglm}
\alias{predict_oomglm}
\title{Internal. Return function that will predict and format output}
\usage{
predict_oomglm(object, type = "response", std_error = FALSE)
}
\arguments{
\item{object}{An \code{oomglm} model object.}

\item{type}{The type of prediction for \code{oomglm} models, "response" or "link".}

\item{std_error}{Indicates if the standard error of predicted means should
be returned.}
}
\description{
Internal. Return function that will predict and format output
}
\keyword{internal}
