% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_oomglm_x}
\alias{predict_oomglm_x}
\title{Internal. Perform \code{oomglm} prediction}
\usage{
predict_oomglm_x(object, new_data, type = "response",
  std_error = FALSE)
}
\arguments{
\item{object}{An \code{oomglm} model object.}

\item{new_data}{Observations for prediction.}

\item{type}{The type of prediction for \code{oomglm} models, "response" or "link".}

\item{std_error}{Indicates if the standard error of predicted means should
be returned.}
}
\description{
Internal. Perform \code{oomglm} prediction
}
\keyword{internal}
