% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oom_predict.R
\name{predict_oomlm}
\alias{predict_oomlm}
\title{return function that will predict and format output}
\usage{
predict_oomlm(object, std_error = FALSE, interval = NULL,
  level = 0.95)
}
\arguments{
\item{object}{\code{oomlm} model object}

\item{std_error}{if TRUE calculate standard error}

\item{interval}{interval type}

\item{level}{the confidence/tolerence for interval calculation}
}
\description{
return function that will predict and format output
}
\keyword{internal}
