% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounded_qr.R
\docType{class}
\name{Rcpp_BoundedQr-class}
\alias{Rcpp_BoundedQr-class}
\title{Class for out of memory QR factorization}
\description{
Class for out of memory QR factorization
}
\details{
Author(s): Alan J. Miller
Source: Journal of the Royal Statistical Society.
Series Applied Statistics, Vol. 41, No. 2
(1992), pp. 458-478
}
\keyword{internal}
