% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals_oomglm}
\alias{residuals_oomglm}
\title{internal wrapper for unpacking data and calculating residuals}
\usage{
residuals_oomglm(object, type = c("deviance", "pearson", "response",
  "working"))
}
\arguments{
\item{object}{\code{oomglm} object}

\item{type}{residual calculation method}
}
\description{
internal wrapper for unpacking data and calculating residuals
}
\keyword{internal}
