% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals_oomglm_x}
\alias{residuals_oomglm_x}
\title{Internal.  Calculate residuals of \code{oomglm} model.}
\usage{
residuals_oomglm_x(object, x, pred, type)
}
\arguments{
\item{object}{An \code{oomglm} model.}

\item{x}{The \code{list} of artifacts returned by \code{unpack_oomchunk()}.}

\item{pred}{predicted response}

\item{`type`}{Residual calculation method.}
}
\description{
Internal.  Calculate residuals of \code{oomglm} model.
}
\keyword{internal}
