% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals_oomlm}
\alias{residuals_oomlm}
\title{Internal. Wrapper for unpacking data and calculating residuals}
\usage{
residuals_oomlm(object)
}
\arguments{
\item{object}{An \code{oomlm} model.}
}
\description{
Internal. Wrapper for unpacking data and calculating residuals
}
\keyword{internal}
