% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oom_residuals.R
\name{residuals_oomlm_x}
\alias{residuals_oomlm_x}
\title{internal \code{oomlm} residual calculation}
\usage{
residuals_oomlm_x(object, x)
}
\arguments{
\item{x}{\code{list} of artifacts returned by \code{unpack_oomchunk()}}
}
\description{
internal \code{oomlm} residual calculation
}
\keyword{internal}
