% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yglm.R
\name{reweight.yglm}
\alias{reweight.yglm}
\title{Reweight \code{yglm} model via Iteratively Reweighted Least Squares method}
\usage{
reweight.yglm(obj, data, num_iter = 1L, tolerance = 1e-07)
}
\arguments{
\item{obj}{\code{yglm} model.}

\item{data}{an \code{oomfeed}, \code{tibble}, \code{dataframe}, \code{list} or \code{environment}.}

\item{num_iter}{number of Fisher scoring iterations to perform.}

\item{tolerance}{tolerance for change in coefficient (as multiple of standard error).}
}
\description{
Reweight \code{yglm} model via Iteratively Reweighted Least Squares method
}
