% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov_biglm.R
\name{rvcov_biglm}
\alias{rvcov_biglm}
\title{vcov implementation for updating linear model
(Miller AS274 QR factorization) as implemented
in package `biglm`}
\usage{
rvcov_biglm(np, D, rbar, ok)
}
\arguments{
\item{np}{number of parameters in model}

\item{D}{diagonals of cross products matrix}

\item{rbar}{the off diagonal portion of the R matrix}
}
\description{
vcov implementation for updating linear model
(Miller AS274 QR factorization) as implemented
in package `biglm`
}
\keyword{internal}
