% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biglm_vcov.R
\name{sandwich_rcov_biglm}
\alias{sandwich_rcov_biglm}
\title{Hubert/White sandwich vcov implementation for updating linear
model (Miller AS274 QR factorization) as implemented 
in package `biglm`}
\usage{
sandwich_rcov_biglm(np, D, rbar, R, betas, ok)
}
\arguments{
\item{np}{number of parameters in model}

\item{D}{diagonals of cross products matrix for sandwich estimator}

\item{rbar}{the off diagonal portion of the R matrix for sandwich estimator}

\item{R}{cov matrix for model}

\item{betas}{model coefficients}

\item{ok}{boolean indicator of linear independence among coefficients}
}
\description{
Hubert/White sandwich vcov implementation for updating linear
model (Miller AS274 QR factorization) as implemented 
in package `biglm`
}
\keyword{internal}
