% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.oomdata}
\alias{tidy.oomdata}
\title{Tidy an \code{oomdata} function}
\usage{
\method{tidy}{oomdata}(x, ...)
}
\arguments{
\item{x}{A \code{oomdata} function.}

\item{...}{Ignored.}
}
\description{
Tidy an \code{oomdata} function
}
