% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_lm.R
\name{update.online_lm}
\alias{update.online_lm}
\title{fit `online_lm` model to new batch of observations}
\usage{
\method{update}{online_lm}(data, obj, ...)
}
\arguments{
\item{data}{an optional data frame, list or environment
(or object coercible by as.data.frame to a data frame)}

\item{object}{online_lm object}
}
\description{
fit `online_lm` model to new batch of observations
}
