% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomlm.R
\name{update.oomlm}
\alias{update.oomlm}
\title{Fit Updating Linear Model to more data}
\usage{
\method{update}{oomlm}(object, data, ...)
}
\arguments{
\item{object}{\code{oomlm} object to be updated.}

\item{data}{an optional \code{oomfeed}, \code{tibble}, \code{dataframe}, \code{list}
or \code{environment}.}

\item{...}{ignored}
}
\description{
Update \code{oomlm} linear model fit with new data.
}
\seealso{
\code{\link[=oomlm]{oomlm()}}
}
