% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R, R/oomglm.R, R/oomlm.R,
%   R/oomlm_robust.R
\name{update}
\alias{update}
\alias{update.oomglm}
\alias{update.oomlm}
\alias{update.oomlm_robust}
\title{Fit Updating Models to more data}
\usage{
\method{update}{oomglm}(object, data, ...)

\method{update}{oomlm}(object, data, ...)

\method{update}{oomlm_robust}(object, data, ...)
}
\arguments{
\item{object}{\code{oomlm} or \code{oomglm} object to be updated.}

\item{data}{an optional \code{oomfeed}, \code{tibble}, \code{dataframe}, \code{list}
or \code{environment}.}

\item{...}{ignored}
}
\description{
Update ploom model fit with new data.
}
\seealso{
\code{\link[=oomlm]{oomlm()}}
}
\keyword{internal}
