% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yglm.R
\name{update.yglm}
\alias{update.yglm}
\title{Update \code{yglm} with new observations}
\usage{
\method{update}{yglm}(obj, data)
}
\arguments{
\item{obj}{\code{yglm} model.}

\item{data}{an \code{oomfeed}, \code{tibble}, \code{dataframe}, \code{list} or \code{environment}.}
}
\description{
Update \code{yglm} with new observations
}
