% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomdata.R
\name{update_oomdata}
\alias{update_oomdata}
\title{Internal. Update \code{oomdata} function record}
\usage{
update_oomdata(obj, chunk_size, complete_iteration)
}
\arguments{
\item{obj}{An \code{oomdata} function.}

\item{chunk_size}{The number of observations in the previous chunk.}

\item{complete_iteration}{Signals that a pass over the data is complete.}
}
\description{
Internal. Update \code{oomdata} function record
}
\details{
Keeps track of number of iterations, chunks, and
observations that have been processed.
}
\keyword{internal}
