% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm.R
\name{update_oomglm}
\alias{update_oomglm}
\title{Internal. Update \code{oomglm} object with additional observations}
\usage{
update_oomglm(object, data, ...)
}
\arguments{
\item{object}{An \code{oomglm} model.}

\item{data}{Observations to fit.}

\item{...}{Ignored.}
}
\description{
This function is typically called via \code{\link[=fit]{fit()}} or \code{\link[=weight]{weight()}}.
}
\keyword{internal}
