% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomlm.R
\name{update_oomlm}
\alias{update_oomlm}
\title{Update Updating Linear Regression Model with more data}
\usage{
update_oomlm(obj, data)
}
\arguments{
\item{obj}{`oomlm` object to be updated}

\item{data}{an optional `oomfeed`, `tibble`, `dataframe`, `list` or `environment`}
}
\description{
Update `oomlm` linear model fit with new data
}
