% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounded_qr.R
\name{vcov.Rcpp_BoundedQr}
\alias{vcov.Rcpp_BoundedQr}
\title{Wrapper for \code{BoundedQr} method \code{vcov}}
\usage{
\method{vcov}{Rcpp_BoundedQr}(object, ...)
}
\arguments{
\item{object}{BoundedQr object}

\item{...}{ignored}
}
\description{
Wrapper for \code{BoundedQr} method \code{vcov}
}
\keyword{internal}
