% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm.R
\name{weight}
\alias{weight}
\title{Fit weighted least squares as part of IRLS iteration}
\usage{
weight(object, data)
}
\arguments{
\item{object}{\code{oomglm} model.}

\item{data}{\code{data.frame} of observations to be fit.}
}
\description{
Fit weighted least squares as part of IRLS iteration
}
