% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yglm.R
\name{yglm}
\alias{yglm}
\title{Initialize Updating Generalized Linear Regression model}
\usage{
yglm(formula, data = NULL, family = gaussian(), weights = NULL,
  start = NULL, ...)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted of class \code{formula}.}

\item{data}{an optional \code{oomfeed}, \code{tibble}, \code{dataframe}, \code{list} or \code{environment}.}

\item{family}{A \code{glm} family object.}

\item{weights}{a one-sided, single term \code{formula} specifying weights.}

\item{start}{starting values for the parameters in the linear predictor.}

\item{...}{ignored.}
}
\description{
Perform  generalized linear regression usig Alan Miller's bounded memory QR
factorization algorithm which enables models with \code{p} variables
to be fit in \code{p^2} memory.
}
