% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{augment.oomglm}
\alias{augment.oomglm}
\title{augment data with prediction, std error, and residuals}
\usage{
\method{augment}{oomglm}(x, data, type = "response", std_error = FALSE,
  ...)
}
\arguments{
\item{x}{\code{oomlm()} model}

\item{data}{\code{tibble()} or other data source}

\item{type}{link or response}

\item{std_error}{calculate standard error of prediction}

\item{...}{ignored}
}
\description{
augment data with prediction, std error, and residuals
}
\seealso{
\code{\link[=augment]{augment()}}, \code{\link[stats:predict.lm]{stats::predict.lm()}}
}
