% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounded_qr.R
\name{coef.Rcpp_BoundedQr}
\alias{coef.Rcpp_BoundedQr}
\title{Wrapper for \code{BoundedQr} method \code{betas}}
\usage{
\method{coef}{Rcpp_BoundedQr}(object, nvar = NULL, ...)
}
\arguments{
\item{object}{BoundedQr object}

\item{nvar}{int number of coefficients to return}

\item{...}{ignored}
}
\description{
Wrapper for \code{BoundedQr} method \code{betas}
}
\keyword{internal}
