% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomlm.R, R/tidy.R
\name{update.oomlm}
\alias{update.oomlm}
\alias{fit.oomlm}
\title{Fit \code{oomlm} model to additional observations}
\usage{
\method{update}{oomlm}(object, data, ...)

\method{fit}{oomlm}(object, data, ...)
}
\arguments{
\item{object}{\code{oomlm} model to be updated}

\item{data}{an optional \code{oomdata_tbl}, \code{oomdata_dbi}, \code{oomdata_con},
\code{tibble}, \code{data.frame}, or \code{list} of observations to fit}

\item{...}{ignored}
}
\description{
Update ploom model fit with new data.
}
\seealso{
\code{\link[=oomlm]{oomlm()}}
}
