% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomlm.R
\name{fit.oomlm}
\alias{fit.oomlm}
\title{Fit \code{oomlm} model to additional observations}
\usage{
\method{fit}{oomlm}(object, data, ...)
}
\arguments{
\item{object}{\code{oomlm} model to be updated.}

\item{data}{An optional \code{oomdata_tbl}, \code{oomdata_dbi}, \code{oomdata_con},
\code{tibble}, \code{data.frame}, or \code{list} of observations to fit.}

\item{...}{Ignored.}
}
\value{
\code{oomlm} object after fitting to \code{data}.
}
\description{
Update ploom model with new data.
}
\seealso{
\code{\link[=oomlm]{oomlm()}}
}
