% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm.R
\name{glm_adjust}
\alias{glm_adjust}
\title{Internal. Apply Generalized Linear Model transformations to \code{chunk}}
\usage{
glm_adjust(object, chunk)
}
\arguments{
\item{object}{An \code{oomglm} model.}

\item{chunk}{List created by \code{unpack_oomchunk()}.}
}
\description{
Internal. Apply Generalized Linear Model transformations to \code{chunk}
}
\keyword{internal}
