% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm.R
\name{glm_adjust}
\alias{glm_adjust}
\title{Apply \code{glm} transformations to \code{chunk}}
\usage{
glm_adjust(object, chunk)
}
\arguments{
\item{object}{\code{oomglm} model.}

\item{chunk}{list created by \code{unpack_oomchunk}.}
}
\description{
Apply \code{glm} transformations to \code{chunk}
}
\keyword{internal}
