% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm.R
\name{init_oomglm}
\alias{init_oomglm}
\title{Internal. Initialize Out of memory Generalized Linear Regression Model}
\usage{
init_oomglm(formula, family, weights, start)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted of class \code{formula}.}

\item{family}{A \code{family} object.}

\item{weights}{A one-sided, single term \code{formula} specifying weights.}

\item{start}{Starting values for the parameters in the linear predictor.}
}
\description{
Internal. Initialize Out of memory Generalized Linear Regression Model
}
\keyword{internal}
