% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm.R
\name{init_oomglm}
\alias{init_oomglm}
\title{Initialize Out of memory Generalized Linear Regression Model}
\usage{
init_oomglm(formula, family, weights, start)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted of class \code{\link[=formula]{formula()}}.}

\item{family}{a \code{\link[=family]{family()}} object.}

\item{weights}{a one-sided, single term \code{\link[=formula]{formula()}} specifying weights.}

\item{start}{starting values for the parameters in the linear predictor.}
}
\description{
Initialize Out of memory Generalized Linear Regression Model
}
\keyword{internal}
