% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomlm.R
\name{init_oomlm}
\alias{init_oomlm}
\title{Internal. Initialize Updating Linear Model}
\usage{
init_oomlm(formula, weights = NULL)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted
of class \code{formula}.}

\item{weights}{A one-sided, single term \code{formula} specifying weights.}
}
\description{
Internal. Initialize Updating Linear Model
}
\keyword{internal}
