% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomlm_robust.R
\name{init_oomlm_robust}
\alias{init_oomlm_robust}
\title{initialize generalized linear model with robust standard errors}
\usage{
init_oomlm_robust(formula, weights = NULL, se_type)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted
of class \code{formula}.}

\item{weights}{A one-sided, single term \code{formula} specifying weights.}
}
\description{
initialize generalized linear model with robust standard errors
}
\keyword{internal}
