% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm.R
\name{init_reweight}
\alias{init_reweight}
\title{Prepare model for IWLS iteration}
\usage{
init_reweight(object)
}
\arguments{
\item{object}{\code{oomglm} model.}
}
\description{
Reset variables used for IWLS calculation
}
