% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm.R
\name{iter_weight}
\alias{iter_weight}
\title{Fit \code{oomglm} model via Iteratively Reweighted Least Squares (IRLS).}
\usage{
iter_weight(object, data, times = 2L, tolerance = 1e-08)
}
\arguments{
\item{object}{\code{oomglm} model.}

\item{data}{an optional \code{oomdata_tbl}, \code{oomdata_dbi}, \code{oomdata_con},
\code{tibble}, \code{data.frame}, or \code{list} of observations to fit}

\item{times}{Maximum number of IRLS iterations to perform. Will
stop iterating if model converges before \code{times} iterations.}

\item{tolerance}{Tolerance used to determine convergence. Represents
change in coefficient as a multiple of standard error.}

\item{...}{ignored}
}
\value{
\code{oomglm} object after performing \code{times} IRLS iterations on
\code{data}.
}
\description{
Fit \code{oomglm} model via Iteratively Reweighted Least Squares (IRLS).
}
\seealso{
\code{\link[=fit]{fit()}}
}
