% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm.R
\name{iter_weight}
\alias{iter_weight}
\title{Fit \code{oomglm} model via Iteratively Reweighted Least Squares (IRLS).}
\usage{
iter_weight(object, data, max_iter = 1L, tolerance = 1e-07)
}
\arguments{
\item{object}{\code{oomglm} model.}

\item{data}{An \code{oom_data}, \code{tibble}, \code{data.frame}, \code{list} or \code{environment}.}

\item{max_iter}{Maximum number of IRLS iterations to perform. Will
stop iterating if model converges before \code{max_iter} iterations.}

\item{tolerance}{Tolerance for change in coefficient as a multiple
of standard error.}
}
\value{
\code{oomglm} object after performing \code{max_iter} IRLS iterations on
\code{data}.
}
\description{
Fit \code{oomglm} model via Iteratively Reweighted Least Squares (IRLS).
}
\seealso{
\code{\link[=oomglm]{oomglm()}}
}
