% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_model.R
\name{model_matrix}
\alias{model_matrix}
\title{lifted from \code{stats}. modified to be faster and less robust}
\usage{
model_matrix(terms, data)
}
\arguments{
\item{data}{data.frame or list}

\item{formula}{\code{stats::terms} object}
}
\description{
lifted from \code{stats}. modified to be faster and less robust
}
\keyword{internal}
