% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_model.R
\name{model_response}
\alias{model_response}
\title{Lifted from \code{stats}. modified to be faster and less robust}
\usage{
model_response(data, type = "any")
}
\arguments{
\item{data}{data.frame or list}

\item{type}{response type}
}
\description{
Lifted from \code{stats}. modified to be faster and less robust
}
\keyword{internal}
