% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm.R
\name{oomglm}
\alias{oomglm}
\alias{predict.oomglm}
\alias{print.oomglm}
\alias{print.summary.oomglm}
\alias{summary.oomglm}
\title{Initialize Updating Generalized Linear model}
\usage{
oomglm(formula, data = NULL, family = gaussian(), weights = NULL,
  start = NULL, ...)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted of class
\link{formula}.}

\item{data}{an optional \link{oom_data}, \link{tibble}, \link{dataframe}, or \link{list}}

\item{family}{a \link{family} object.}

\item{weights}{a one-sided, single term \link{formula} specifying weights.}

\item{start}{starting values for the parameters in the linear predictor.}

\item{...}{ignored.}
}
\value{
It is up to the user to know when fitting is complete.
Therefore, only basic model characteristics are provided as values with
the \code{oomglm} object. Statistics are available on demand via \code{summary} and
extractor functions.

\item{converged}{Indicates if the IRLS algorithm has converged.}
\item{iter}{The number of iterations of IRLS performed.}
\item{n}{The number observations processed per round of IRLS.}
\item{df.residual}{The residual degrees of freedom.}
\item{df.null}{The residual degrees of freedom.}
\item{formula}{the \code{\link[=formula]{formula()}} object specifying the linear model.}
\item{family}{a \code{\link[=family]{family()}} object describing the error distribution
and link function used in the model.}
\item{terms}{The \code{\link[=terms]{terms()}} object used.}
\item{weights}{The weights \code{\link[=formula]{formula()}} provided to the model.}
\item{call}{The matched call.}
}
\description{
Perform  generalized linear regression using Alan Miller's bounded memory QR
factorization algorithm which enables models with \code{p} variables
to be fit in \code{p^2} memory.
}
\details{
\code{oomglm()} initializes an object of class \code{oomglm} inheriting from the
class \code{oomlm}. \code{oomglm} objects are fit via iteratively IRLS
Iterative fitting is performed with the function \code{\link[=iter_weight]{iter_weight()}}.

A \code{oomglm} object can be in various states of fit depending on the number
of seen observations and rounds of IRLS that have been performed.
It is important to view the model within the context of:
the number of observations processed per round of IRLS (\code{n});
the number of IRLS iterations that have been performed (\code{iter});
and if the IRLS algorithm has converged (\code{converged}).
}
\examples{
\donttest{
# The `oomglm()` function employs Iteratively Weighted Least Squares (IWLS).
# The IWLS iterations are performed by the function `iter_weight()` which
# makes passes over the data until estimate convergence.

# re-weight 4 times or until convergence
x <- oomglm(mpg ~ cyl + disp)
x <- iter_weight(x, mtcars, max_iter = 4)

tidy(x)

# To fit data in chunks, use `oom_data()`:

y      <- oomglm(mpg ~ cyl + disp)
chunks <- oom_data(mtcars, chunk_size = 10)
y      <- iter_weight(x, chunks, max_iter = 4)

tidy(y)

}
}
\seealso{
\code{\link[=oomlm]{oomlm()}}
}
