% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm.R
\name{oomglm}
\alias{oomglm}
\alias{print.oomglm}
\alias{print.summary.oomglm}
\alias{summary.oomglm}
\alias{logLik.oomglm}
\alias{family.oomglm}
\alias{deviance.oomglm}
\alias{vcov.oomglm}
\alias{AIC.oomglm}
\title{Out of memory Generalized Linear model}
\usage{
oomglm(formula, family = gaussian(), weights = NULL, start = NULL,
  ...)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted of class
\code{formula}.}

\item{family}{A \code{family} object.}

\item{weights}{A one-sided, single term \code{formula} specifying weights.}

\item{start}{Starting values for the parameters in the linear predictor.}

\item{...}{Ignored.}
}
\value{
It is up to the user to know when fitting is complete.
Therefore, only basic model characteristics are provided as values.
Statistics are available on demand via summary and extractor functions.

\item{converged}{Indicates if the IRLS algorithm has converged.}
\item{iter}{The number of iterations of IRLS performed.}
\item{n}{The number observations processed per round of IRLS.}
\item{df.residual}{The residual degrees of freedom.}
\item{df.null}{The residual degrees of freedom.}
\item{formula}{the \code{formula} object specifying the linear model.}
\item{family}{a \code{family} object describing the error distribution
and link function used in the model.}
\item{terms}{The \code{terms} object used.}
\item{weights}{The weights \code{formula} provided to the model.}
\item{call}{The matched call.}
}
\description{
Perform memory-efficient generalized linear regression using
the AS274 bounded memory QR factorization algorithm.
}
\details{
An \code{oomglm} model can be in various states of fit depending on the number
of seen observations and rounds of IRLS that have been performed.
It is important to view the model within the context of:
the number of observations processed per round of IRLS (\code{n});
the number of IRLS iterations that have been performed (\code{iter});
and if the IRLS algorithm has converged (\code{converged}).
}
\examples{
\donttest{
# The `oomglm()` function employs Iteratively Weighted Least Squares (IRLS).
# The IRLS iterations are performed by the function `fit()` which
# makes passes over the data until convergence.

# reweight 4 times or until convergence
x <- oomglm(mpg ~ cyl + disp)
x <- fit(x, mtcars, times = 4)

tidy(x)

# To fit in-memory data in chunks, use `oomdata_tbl()`:

y      <- oomglm(mpg ~ cyl + disp)
chunks <- oomdata_tbl(mtcars, chunk_size = 10)
y      <- fit(y, chunks, times = 4)

tidy(y)

}
}
\seealso{
\code{\link[=oomlm]{oomlm()}}
}
