% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm_robust.R
\name{oomglm_robust}
\alias{oomglm_robust}
\alias{vcov.oomglm_robust}
\title{Out of Memory Generalized Linear model with robust standard errors}
\usage{
oomglm_robust(formula, data = NULL, family = gaussian(),
  weights = NULL, start = NULL, se_type = "H1", ...)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted of class
\code{formula}.}

\item{data}{an optional \code{oom_data}, \code{tibble}, \code{data.frame}, \code{list} or
\code{environment}.}

\item{family}{A \code{glm} family object.}

\item{weights}{a one-sided, single term \code{formula} specifying weights.}

\item{start}{starting values for the parameters in the linear predictor.}

\item{se_type}{string indicating what se type to usecan be "HC0", "HC1"
"stata", or "classical".  default "HC1".}

\item{...}{ignored.}
}
\value{
It is up to the user to know when fitting is complete.
Therefore, only basic model characteristics are provided as values with
the \code{oomglm} object. Statistics are available on demand via \code{summary} and
extractor functions.

\item{converged}{Indicates if the IRLS algorithm has converged.}
\item{iter}{The number of iterations of IRLS performed.}
\item{n}{The number observations processed per round of IRLS.}
\item{df.residual}{The residual degrees of freedom.}
\item{df.null}{The residual degrees of freedom.}
\item{formula}{the \code{\link[=formula]{formula()}} object specifying the linear model.}
\item{family}{a \code{\link[=family]{family()}} object describing the error distribution
and link function used in the model.}
\item{terms}{The \code{\link[=terms]{terms()}} object used.}
\item{weights}{The weights \code{\link[=formula]{formula()}} provided to the model.}
\item{call}{The matched call.}
}
\description{
Perform linear regression via Alan Miller's bounded memory QR
factorization algorithm enables models with \code{p} variables with
\code{p} variables  to be fit with robust standard errors in \code{p^4} memory.
}
\details{
\code{oomglm_robust} initializes an object of class \code{oomlm}. \code{oomlm} objects
are intended to be iteratively updated with new data via the function
\code{\link[=update]{update()}}. If \code{data} is provided to the \code{oolm} function call, an
\code{\link[=update]{update()}} round will be performed on initialization.

The provided \code{formula} must not contain any data-dependent terms to ensure
consistency across calls to \code{\link[=update]{update()}}. Factors are permitted, but the
levels of the factor must be the same across all data chunks. Empty factor
levels are accepted.
}
\seealso{
\code{\link[=oomlm_robust]{oomlm_robust()}}
}
