% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomlm_robust.R
\name{oomlm_robust}
\alias{oomlm_robust}
\alias{vcov.oomlm_robust}
\title{Out of memory Linear model with robust standard errors}
\usage{
oomlm_robust(formula, weights = NULL, se_type = "HC1", ...)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted of class
\code{formula}.}

\item{weights}{A one-sided, single term \code{formula} specifying weights.}

\item{se_type}{Indicates what Standard Error method to use: "HC0", "HC1"
"stata", or "classical".}

\item{...}{Ignored.}
}
\value{
A \code{oomlm_robust} model is perpetually in an \emph{in-progress} state. It is up
to the user to know when fitting is complete. Therefore, only basic
model characteristics are provided as values. Statistics are available on
demand via summary and extractor functions.

\item{n}{The number of observations processed.}
\item{df.residual}{The residual degrees of freedom.}
\item{formula}{The \code{formula} object specifying the linear model.}
\item{terms}{The \code{terms} object specifying the terms of the linear model.}
\item{weights}{a one-sided, single term \code{formula} specifying weights.}
\item{call}{The matched call.}
}
\description{
Perform memory-efficient generalized linear regression using
the AS274 bounded memory QR factorization algorithm and estimate
robust standard errors.
}
\details{
The provided \code{formula} must not contain any data-dependent terms to ensure
consistency across calls to \code{fit()}. Factors are permitted, but the
levels of the factor must be the same across all data chunks. Empty factor
levels are accepted.
}
\seealso{
\code{\link[=oomlm]{oomlm()}}
}
