% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oom_predict.R
\name{predict}
\alias{predict}
\alias{predict.oomlm}
\title{Predicted values based on \code{oomlm} object}
\usage{
\method{predict}{oomlm}(object, data = NULL, se_fit = FALSE,
  as_function = FALSE, ...)
}
\arguments{
\item{object}{object inheriting from class \code{oomlm}}

\item{data}{observations for prediction}

\item{se_fit}{indicates if the standard error of predicted means should
be returned}

\item{as_function}{if TRUE a function with only a \code{data} argument is returned
for subsequent fitting}

\item{...}{ignored}
}
\description{
Predicted values based on \code{oomlm} object
}
