% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oom_predict.R
\name{predict}
\alias{predict}
\alias{predict.oomlm}
\alias{predict.oomglm}
\title{Predict values using \code{oomlm()} and \code{oomglm()} models}
\usage{
\method{predict}{oomlm}(object, new_data = NULL, se_fit = FALSE,
  interval = NULL, level = 0.95, as_function = FALSE, ...)

\method{predict}{oomglm}(object, new_data, type = c("link", "response"),
  se_fit = FALSE, as_function = FALSE, ...)
}
\arguments{
\item{object}{object inheriting from class \code{oomlm}}

\item{new_data}{observations for prediction}

\item{se_fit}{indicates if the standard error of predicted means should
be returned}

\item{interval}{type of interval calculation}

\item{level}{tolerance/confidence level}

\item{as_function}{if TRUE a function with only a \code{data} argument is returned
for subsequent fitting}

\item{...}{ignored}

\item{type}{the type of prediction, one of "reponse" or "link"}
}
\description{
Predict values using \code{oomlm()} and \code{oomglm()} models
}
\examples{
\donttest{
# fit an `oomlm` model
chunks <- oomdata_tbl(mtcars, chunk_size = 1)
x  <- fit(oomlm(mpg ~ cyl + disp + hp), chunks)

# call `predict`
pred <- predict(x, mtcars)
sum((pred - mtcars$mpg)^2)

# pass TRUE for the `as_function` argument and the
# return value will be a prediction function with
# only one argument for data
pred_fun <- predict(x, mtcars, as_function = TRUE)
rss      <- 0
while(!is.null(chunk <- chunks())) {
 rss <- rss + (pred_fun(chunk) - chunk[, "mpg"])^2
}
rss

# pass TRUE for the `se_fit` argument and the
# return value will include standard errors
# for the predicted means
pred <- predict(x, mtcars, se_fit = TRUE)
names(pred)
head(pred$se)

}
}
