% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\alias{predict.oomlm}
\alias{predict.oomglm}
\title{Predict values using \code{oomlm} and \code{oomglm} models}
\usage{
\method{predict}{oomlm}(object, new_data = NULL, std_error = FALSE,
  interval = NULL, level = 0.95, as_function = FALSE, ...)

\method{predict}{oomglm}(object, new_data = NULL, type = "response",
  std_error = FALSE, as_function = FALSE, ...)
}
\arguments{
\item{object}{An object inheriting from class \code{oomlm}.}

\item{new_data}{Observations for prediction.}

\item{std_error}{Indicates if the standard error of predicted means should
be returned.}

\item{interval}{Type of interval calculation, "confidence" or "prediction.
Is ignored if \code{std_error} is \code{FALSE}.}

\item{level}{Confidence level for interval calculation.}

\item{as_function}{If \code{TRUE}, a function requiring only \code{new_data} is
returned for subsequent fitting.}

\item{...}{Ignored.}

\item{type}{The type of prediction for \code{oomglm} models, "response" or "link".}
}
\description{
Predict values using \code{oomlm} and \code{oomglm} models
}
\examples{
\donttest{
# fit an `oomlm` model
chunks <- oomdata_tbl(mtcars, chunk_size = 1)
x  <- fit(oomlm(mpg ~ cyl + disp + hp), chunks)

# call `predict`
pred <- predict(x, mtcars)
sum((pred - mtcars$mpg)^2)

# pass TRUE for the `as_function` argument and the
# return value will be a prediction function with
# only one argument for data
pred_fun <- predict(x, mtcars, as_function = TRUE)
pred_fun(mtcars[1, ])

# pass TRUE for the `std_error` argument and the
# return value will include standard errors
# for the predicted means
pred <- predict(x, mtcars, std_error = TRUE)
head(pred)

# with std_error true, we can specify that a confidence or
# prediction interval be returned
pred <- predict(x, mtcars, std_error = TRUE, interval = "confidence")
head(pred)
}

}
\seealso{
\code{\link[=oomlm]{oomlm()}}, \code{\link[=oomglm]{oomglm()}}
}
