% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oom_predict.R
\name{predict_oomglm}
\alias{predict_oomglm}
\title{return function that will predict and format output}
\usage{
predict_oomglm(object, type = "response", std_error = FALSE)
}
\arguments{
\item{object}{\code{oomglm} model object}

\item{type}{type of prediction, link or response}

\item{std_error}{if TRUE calculate standard error}
}
\description{
return function that will predict and format output
}
\keyword{internal}
