% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_oomglm_x}
\alias{predict_oomglm_x}
\title{perform \code{oomglm} prediction}
\usage{
predict_oomglm_x(object, new_data, type = "response",
  std_error = FALSE)
}
\arguments{
\item{object}{\code{oomglm} model object}

\item{new_data}{covariates for prediction}

\item{type}{type of prediction, link or response}

\item{std_error}{if TRUE calculate standard error}
}
\description{
perform \code{oomglm} prediction
}
\keyword{internal}
