% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_oomlm}
\alias{predict_oomlm}
\title{Internal. Return function that will predict and format output}
\usage{
predict_oomlm(object, std_error = FALSE, interval = NULL,
  level = 0.95)
}
\arguments{
\item{object}{An \code{oomlm} model object.}

\item{std_error}{Indicates if the standard error of predicted means should
be returned.}

\item{interval}{Type of interval calculation, "confidence" or "prediction.
Is ignored if \code{std_error} is FALSE.}

\item{level}{Confidence level for interval calculation.}
}
\description{
Internal. Return function that will predict and format output
}
\keyword{internal}
