% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_oomlm_x}
\alias{predict_oomlm_x}
\title{Internal. Perform \code{oomlm} prediction}
\usage{
predict_oomlm_x(object, new_data, std_error = FALSE, interval = NULL,
  level = 0.95)
}
\arguments{
\item{object}{An \code{oomlm} model object.}

\item{new_data}{Observations for prediction.}

\item{std_error}{Indicates if the standard error of predicted means should
be returned.}

\item{interval}{Type of interval calculation, "confidence" or "prediction.
Is ignored if \code{std_error} is FALSE.}

\item{level}{Confidence level for interval calculation.}
}
\description{
Internal. Perform \code{oomlm} prediction
}
\keyword{internal}
