% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oom_predict.R
\name{predict_oomlm_x}
\alias{predict_oomlm_x}
\title{perform \code{oomlm} prediction}
\usage{
predict_oomlm_x(object, new_data, std_error = FALSE, interval = NULL,
  level = 0.95)
}
\arguments{
\item{object}{\code{oomlm} model object}

\item{new_data}{covariates for prediction}

\item{std_error}{if TRUE calculate standard error}

\item{interval}{interval type}

\item{level}{the confidence/tolerence for interval calculation}
}
\description{
perform \code{oomlm} prediction
}
\keyword{internal}
