% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oom_residuals.R
\name{resid}
\alias{resid}
\alias{resid.oomlm}
\alias{resid.oomglm}
\title{calculate residuals for oomglm}
\usage{
\method{resid}{oomlm}(object, data, as_function = FALSE, ...)

\method{resid}{oomglm}(object, data, type = c("deviance", "pearson",
  "response", "working"), as_function = FALSE, ...)
}
\arguments{
\item{object}{\code{oomlm} or \code{oomglm} model}

\item{data}{data to calculate residuals}

\item{as_function}{if TRUE a function with only a \code{data} argument is returned
for subsequent residual calculations}

\item{...}{ignored}

\item{type}{residual calculation method}
}
\description{
calculate residuals for oomglm

#' @rdname resid
resid     <- function (x, ...) { UseMethod("resid", x) }
#' @rdname resid
residuals <- function (x, ...) { UseMethod("resid", x) }
}
