% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oom_residuals.R
\name{residuals}
\alias{residuals}
\alias{resid.oomlm}
\alias{residuals.oomlm}
\alias{resid.oomglm}
\alias{residuals.oomglm}
\title{calculate residuals for oomglm}
\usage{
\method{residuals}{oomlm}(object, data, as_function = FALSE, ...)

\method{residuals}{oomlm}(object, data, as_function = FALSE, ...)

\method{residuals}{oomglm}(object, data, type = c("deviance", "pearson",
  "response", "working"), as_function = FALSE, ...)

\method{residuals}{oomglm}(object, data, type = c("deviance", "pearson",
  "response", "working"), as_function = FALSE, ...)
}
\arguments{
\item{object}{\code{oomlm} or \code{oomglm} model}

\item{data}{data to calculate residuals}

\item{as_function}{if TRUE a function with only a \code{data} argument is returned
for subsequent residual calculations}

\item{...}{ignored}

\item{type}{residual calculation method}
}
\description{
calculate residuals for oomglm
}
