% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals_oomglm}
\alias{residuals_oomglm}
\title{Internal. Wrapper for unpacking data and calculating residuals}
\usage{
residuals_oomglm(object, type = c("deviance", "pearson", "response",
  "working"))
}
\arguments{
\item{object}{An \code{oomglm} model.}

\item{`type`}{Residual calculation method.}
}
\description{
Internal. Wrapper for unpacking data and calculating residuals
}
\keyword{internal}
