% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals_oomglm_x}
\alias{residuals_oomglm_x}
\title{internal \code{oomglm} residual calculation}
\usage{
residuals_oomglm_x(object, x, type)
}
\arguments{
\item{x}{\code{list} of artifacts returned by \code{unpack_oomchunk()}}
}
\description{
internal \code{oomglm} residual calculation
}
\keyword{internal}
