% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals_oomlm}
\alias{residuals_oomlm}
\title{internal wrapper for unpacking data and calculating residuals}
\usage{
residuals_oomlm(object)
}
\arguments{
\item{object}{\code{oomlm} object}
}
\description{
internal wrapper for unpacking data and calculating residuals
}
\keyword{internal}
