% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals_oomlm_x}
\alias{residuals_oomlm_x}
\title{Internal.  Calculate residuals of \code{oomlm} model}
\usage{
residuals_oomlm_x(object, x)
}
\arguments{
\item{object}{An \code{oomlm} model.}

\item{x}{The \code{list} of artifacts returned by \code{unpack_oomchunk()}.}
}
\description{
Internal.  Calculate residuals of \code{oomlm} model
}
\keyword{internal}
