% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm.R
\name{reweight}
\alias{reweight}
\title{Reweight \code{oomglm} model via Iteratively Reweighted Least Squares (IWLS).}
\usage{
reweight(object, data, max_iter, tolerance = 1e-07)
}
\arguments{
\item{object}{\code{oomglm} model.}

\item{data}{An \code{oomfeed}, \code{tibble}, \code{dataframe}, \code{list} or \code{environment}.}

\item{max_iter}{Maximum number of IWLS iterations to perform. Will
stop iterating if model converges before \code{max_iter} iterations.}

\item{tolerance}{Tolerance for change in coefficient as a multiple
of standard error.}
}
\value{
\code{oomglm} object after performing \code{max_iter} IWLS iterations on
\code{data}.
}
\description{
Reweight \code{oomglm} model via Iteratively Reweighted Least Squares (IWLS).
}
\seealso{
\code{\link[=oomglm]{oomglm()}}
}
