% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm.R
\name{update.oomglm}
\alias{update.oomglm}
\title{Update \code{oomglm} with new observations}
\usage{
\method{update}{oomglm}(obj, data)
}
\arguments{
\item{obj}{\code{oomglm} model.}

\item{data}{an \code{oomfeed}, \code{tibble}, \code{dataframe}, \code{list} or \code{environment}.}
}
\description{
Update \code{oomglm} with new observations
}
