% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounded_qr.R
\name{update.Rcpp_BoundedQr}
\alias{update.Rcpp_BoundedQr}
\title{Wrapper for \code{BoundedQr} method \code{update}}
\usage{
\method{update}{Rcpp_BoundedQr}(object, X, y, weights, ...)
}
\arguments{
\item{object}{BoundedQr object to update}

\item{X}{numeric matrix of covariate observations}

\item{y}{numeric vector of response}

\item{weights}{numeric vector of observation weights}

\item{...}{ignored}
}
\description{
Wrapper for \code{BoundedQr} method \code{update}
}
\keyword{internal}
