% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ylm.R
\name{update.ylm}
\alias{update.ylm}
\title{Update Updating Linear Regression Model with more data}
\usage{
\method{update}{ylm}(obj, data)
}
\arguments{
\item{obj}{`ylm` object to be updated}

\item{data}{an optional `oomfeed`, `tibble`, `dataframe`, `list` or `environment`}
}
\description{
Update `ylm` linear model fit with new data
}
