% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomglm.R
\name{update_oomglm}
\alias{update_oomglm}
\title{Update \code{\link[=oomglm]{oomglm()}} object with additional observations}
\usage{
update_oomglm(object, data, ...)
}
\arguments{
\item{object}{\code{\link[=oomglm]{oomglm()}} model}

\item{data}{observations to fit}

\item{...}{ignored}
}
\description{
This function is typically called via \code{\link[=iter_weight]{iter_weight()}} or \code{\link[=weight]{weight()}}.
It is exposed for potential non-typical situations.
}
\keyword{internal}
