% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomlm.R
\name{update_oomlm}
\alias{update_oomlm}
\title{fit `oomlm` model to new batch of observations}
\usage{
update_oomlm(obj, data)
}
\arguments{
\item{data}{an optional data frame, list or environment
(or object coercible by as.data.frame to a data frame)}

\item{oomlm}{model object}
}
\description{
fit `oomlm` model to new batch of observations
}
