% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounded_qr.R
\name{vcov.Rcpp_BoundedQr}
\alias{vcov.Rcpp_BoundedQr}
\title{wrapper for `BoundedQr` method `vcov`}
\usage{
\method{vcov}{Rcpp_BoundedQr}(qr)
}
\arguments{
\item{qr}{BoundedQr object}
}
\description{
wrapper for `BoundedQr` method `vcov`
}
\keyword{internal}
